requirements.txt
streamlit>=1.31.0
langchain>=0.1.0
langchain-community>=0.0.10
python-dotenv>=1.0.0
pydantic>=2.0.0
openai>=1.10.0
watchdog==3.0.0 